requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.3
email-validator==2.1.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
schedule==1.2.0
pytest==7.4.3
